investigators should die and be transported to the hospital or become mad and be interned in asylum.

+++ app/models/ga/investigate.rb
@@ -10,8 +10,16 @@ class Ga::Investigate < Ga::AskPeople
         if clue
           # On a successful observation test, the investigator get the indice
           ActiveRecord::Base.transaction do
-            #  First we remove this clue
+
+            # Mark the clue as found in the news
+            in_the_news= InTheNewsPlace.find_by_game_board_id_and_place_id( clue.game_board_id, clue.place_id )
+            raise "#{self.class}##{__method__} : Unable to find News for clue #{clue.inspect}" unless in_the_news
+            in_the_news.update_attribute( :discovered_clue, true )
+
+            # Then remove the clue
             clue.destroy
+
+            # Add the clue to investigator pool
             investigator.update_attribute( :clues, investigator[ :clues ] + 1 )
           end
           @action_result = :find_clue
